# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for bulk import/export of users, articles, and comments with streaming and async processing",
        "title": "Bulk Import/Export API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/exports": {
            "get": {
                "description": "Streams users, articles, or comments directly in CSV or NDJSON format",
                "produces": [
                    "text/csv",
                    "application/x-ndjson"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Stream export data (synchronous)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource type (users, articles, or comments)",
                        "name": "resource",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Export format (csv or ndjson)",
                        "name": "format",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Streaming export data",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates an export job to export filtered data asynchronously with download URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Create async export job",
                "parameters": [
                    {
                        "description": "Export configuration",
                        "name": "export",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/exports.CreateExportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Existing job returned (idempotency)",
                        "schema": {
                            "$ref": "#/definitions/exports.CreateExportResponse"
                        }
                    },
                    "202": {
                        "description": "Export job created",
                        "schema": {
                            "$ref": "#/definitions/exports.CreateExportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/exports/{job_id}": {
            "get": {
                "description": "Retrieves the status and download URL of an export job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports"
                ],
                "summary": "Get export job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Export Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Export job details with download URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "exports.CreateExportRequest": {
            "type": "object",
            "required": [
                "format",
                "idempotency_key",
                "resource_type"
            ],
            "properties": {
                "fields": {
                    "description": "Optional field selection",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filters": {
                    "description": "Optional filters",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "format": {
                    "type": "string",
                    "enum": [
                        "csv",
                        "ndjson"
                    ]
                },
                "idempotency_key": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string",
                    "enum": [
                        "users",
                        "articles",
                        "comments"
                    ]
                }
            }
        },
        "exports.CreateExportResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}